backend/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.8.0
orjson==3.10.3
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
